Finished all teacher dashboard components.
