+++ src/App.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { Redirect, Route, Switch } from 'react-router-dom';
 import {
   CookiePopup,
-  Header,
   LandingPageWrapper,
   LoadUserData,
   LogoutPopup,
@@ -24,7 +23,6 @@ const App = (): React.ReactElement => {
       <SEO />
       <CookiePopup />
       <LogoutPopup />
-      <PrivateRoute component={Header} />
       <Switch>
         {/* Public Routes */}
         <Route
