--- src/components/common/SectionList/index.ts
@@ -1,1 +1,0 @@
-export { default as SectionList } from './SectionList';
