+++ src/components/common/index.ts
@@ -9,6 +9,5 @@ export { LoadUserData } from './LoadUserData';
 export { LogoutPopup } from './LogoutPopup';
 export { Modal } from './Modal';
 export { PrivateRoute } from './PrivateRoute';
-export { SectionList } from './SectionList';
 export * as Select from './Select';
 export { SEO } from './SEO';
