+++ src/components/pages/TeacherDashboard/TeacherDashboardContainer.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { Redirect, Route, Switch } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import { sections } from '../../../state';
+import { Header } from '../../common';
 import { CreateNewRumble } from './CreateNewRumble';
 import RenderTeacherDashboard from './RenderTeacherDashboard';
 import { TeacherViewRumble } from './TeacherViewRumble';
@@ -13,43 +14,46 @@ const TeacherDashboardContainer = (): React.ReactElement => {
   const sectionList = useRecoilValue(sections.list);
 
   return (
-    <Switch>
-      <Route
-        exact
-        path="/dashboard/teacher"
-        render={() => <RenderTeacherDashboard sectionList={sectionList} />}
-      />
-      <Route
-        exact
-        path="/dashboard/teacher/section"
-        render={() => <TeacherViewSection />}
-      />
-      <Route
-        exact
-        path="/dashboard/teacher/rumble"
-        render={() => <TeacherViewRumble />}
-      />
-      <Route
-        exact
-        path="/dashboard/teacher/student"
-        render={() => <TeacherViewStudent />}
-      />
-      <Route
-        exact
-        path="/dashboard/teacher/submission"
-        render={() => <TeacherViewSubmission />}
-      />
-      <Route
-        exact
-        path="/dashboard/teacher/rumble/new"
-        render={(props) => <CreateNewRumble {...props} />}
-      />
-      {/* Fallback redirect for nonexistent routes */}
-      <Route
-        path="/dashboard/teacher"
-        component={() => <Redirect to="/dashboard/teacher" />}
-      />
-    </Switch>
+    <>
+      <Header />
+      <Switch>
+        <Route
+          exact
+          path="/dashboard/teacher"
+          render={() => <RenderTeacherDashboard sectionList={sectionList} />}
+        />
+        <Route
+          exact
+          path="/dashboard/teacher/section"
+          render={() => <TeacherViewSection />}
+        />
+        <Route
+          exact
+          path="/dashboard/teacher/rumble"
+          render={() => <TeacherViewRumble />}
+        />
+        <Route
+          exact
+          path="/dashboard/teacher/student"
+          render={() => <TeacherViewStudent />}
+        />
+        <Route
+          exact
+          path="/dashboard/teacher/submission"
+          render={() => <TeacherViewSubmission />}
+        />
+        <Route
+          exact
+          path="/dashboard/teacher/rumble/new"
+          render={(props) => <CreateNewRumble {...props} />}
+        />
+        {/* Fallback redirect for nonexistent routes */}
+        <Route
+          path="/dashboard/teacher"
+          component={() => <Redirect to="/dashboard/teacher" />}
+        />
+      </Switch>
+    </>
   );
 };
