+++ src/components/pages/TeacherDashboard/TeacherDashboardRumbleList/TeacherDashboardRumbleCard.tsx
@@ -4,7 +4,7 @@ import { useSetRecoilState } from 'recoil';
 import { Rumbles, Sections } from '../../../../api';
 import { current } from '../../../../state';
 
-const TeacherRumble = ({
+const TeacherDashboardRumbleCard = ({
   section,
   rumble,
 }: ITeacherDashboardRumbleCardProps): React.ReactElement => {
@@ -20,11 +20,18 @@ const TeacherRumble = ({
 
   return (
     <div className="rumble-card" onClick={openCurrentRumble}>
-      <h3>{rumble.sectionName}</h3>
-      <p>ID: {rumble.id}</p>
-      <p>SectionId: {rumble.sectionId}</p>
-      <p>PromptID: {rumble.promptId}</p>
-      <p>This rumble is {rumble.numMinutes} minutes long!</p>
+      <div className="content">
+        <h3>Class Name</h3>
+        <h4>{rumble.sectionName}</h4>
+      </div>
+      <div className="content">
+        <h3>Status</h3>
+        <h4>{rumble.end_time ? 'Active' : 'Scheduled'}</h4>
+      </div>
+      <div className="content">
+        <h3>Length</h3>
+        <h4>{rumble.numMinutes} minutes</h4>
+      </div>
     </div>
   );
 };
@@ -34,4 +41,4 @@ interface ITeacherDashboardRumbleCardProps {
   rumble: Rumbles.IRumbleWithSectionInfo;
 }
 
-export default TeacherRumble;
+export default TeacherDashboardRumbleCard;
