+++ src/components/pages/TeacherDashboard/TeacherDashboardSectionList/index.ts
@@ -1,0 +1,1 @@
+export { default as TeacherDashboardSectionList } from './TeacherDashboardSectionList';
