+++ src/styles/components/pages/teacherDashboard/promptQueueDisplay.scss
@@ -14,33 +14,10 @@
       margin-bottom: 1.5rem;
     }
     .prompt-queue-list {
-      display: flex;
-      flex-flow: row nowrap;
-      overflow-x: scroll;
-      overflow-y: hidden;
-      @include bp(mid) {
-        overflow: visible;
-        flex-flow: row wrap;
-        padding-bottom: unset;
-      }
-
+      @include scaling-flexbox();
       div {
-        padding: 1.6rem;
-        border-radius: 1.6rem;
-        margin-bottom: 2rem;
-        margin-right: 2rem;
-
-        min-width: 20rem;
-        @include bp(mid) {
-          min-width: unset;
-          width: 45%;
-        }
-        @include bp(750px) {
-          width: 30%;
-        }
-
+        @include card;
         &.prompt-queue-item {
-          @include clickable-bg($white, 2.5%);
           border: 1px solid $c-grey-40;
           h3 {
             color: $c-grey-40;
@@ -48,13 +25,13 @@
             font-weight: 800;
             padding-bottom: 0.7rem;
             border-bottom: 1px solid $c-grey-20;
-            margin-bottom: 1rem;
             text-transform: uppercase;
           }
           p {
             color: $black;
             font-size: 1.6rem;
             line-height: 2rem;
+            margin: 1.6rem 0 1rem;
           }
         }
