+++ src/styles/components/pages/teacherDashboard/teacherDashboard.scss
@@ -7,3 +7,5 @@
 @import './studentList.scss';
 @import './promptQueueDisplay.scss';
 @import './submissionList.scss';
+@import './teacherDashboardRumbleList.scss';
+@import './teacherDashboardSectionList.scss';
