+++ src/styles/globals.scss
@@ -38,6 +38,8 @@ $primary-lite: #d5e6fb;
 $primary-base: #5872ff;
 $primary-dark: #093269;
 
+$slate: #8f9bb3;
+
 // Testing different error colors
 $red: #a30015;
 $red-bright: #e0223b;
@@ -50,7 +52,6 @@ $w-grey-50: #717171;
 $w-grey-60: #535353;
 $w-grey-90: #262626;
 
-$c-grey-base: #8f9bb3;
 $c-grey-20: #cbd2d9;
 $c-grey-40: #7b8794;
 $c-grey-70: #47525d;
